requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
faiss-cpu==1.10.0
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pymongo==4.11.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
